requirements.txt
fastapi==0.135.3
fastapi[standard]==0.135.3
sqlalchemy==2.0.49
sqladmin==0.24.0
asyncpg==0.31.0
alembic==1.18.4
# pydantic==2.12.5
pydantic_settings==2.13.1
faststream==0.6.7
faststream[cli]==0.6.7
faststream[redis]==0.6.7
pyjwt==2.12.1
pwdlib[argon2]==0.3.0
